Load images as gif indexed images

+++ src/sig/engine/GameEntity.java
@@ -2,5 +2,5 @@ package sig.engine;
 
 public interface GameEntity {
     public void update(double updateMult);
-    public void draw(int[] p);
+    public void draw(byte[] p);
 }
